fix some dtype mismatch problems in loss function

+++ metrics.py
@@ -29,7 +29,7 @@ class KTLoss(nn.Module):
         pred_answers = torch.cat(((1.0 - pred_answers).unsqueeze(dim=1), pred_answers.unsqueeze(dim=1)), dim=1)
         # pred_answers shape: [batch_size, 2, seq_len - 1]
         nll_loss = nn.NLLLoss(ignore_index=-1)  # ignore masked values in real_answers
-        loss = nll_loss(pred_answers, real_answers)
+        loss = nll_loss(pred_answers, real_answers.long())
         return loss
